Add files via upload
made two pipelines one with fetch and one with selenium. So basically when there is a get request of urls from frontend. The urls are separated into two types one which are of linkedin or twitter or reddit that is social media urls and one with other websites that is other urls. The social media urls are scraped with selenium pipeline using its drivers and proxies are put in chrome options and other urls are send to scrape from fetch pipeline.